Bring back the devicedId field.
There was one vulnerable in the previous version that it will automatically connect to next user in firestore if user doesnt exist anymore in the database.
The mistake is that misunderstand that deviceId doesnt important anymore.
One more thing to add later: Log out feature, like doing unpair the lineUserId with deviceId. then change the line account to pair with.